--- a/cli/main.rs
+++ b/cli/main.rs
@@ -1,8 +1,5 @@
-use core::{fmt, str::FromStr};
-use exitfailure::ExitFailure;
-use failure::{format_err, Error};
+use anyhow::{anyhow, Error, Result};
 use multibase::Base;
-use std::io::{self, Read, Write};
 use structopt::StructOpt;
 
 #[derive(StructOpt, Debug)]
@@ -17,65 +14,87 @@ enum Mode {
     #[structopt(name = "encode")]
     Encode {
         /// The base to use for encoding.
-        #[structopt(short = "b", long = "base", default_value = "base58-btc")]
+        #[structopt(short = "b", long = "base", default_value = "base58btc")]
         base: StrBase,
+        /// The data need to be encoded.
+        #[structopt(short = "i", long = "input")]
+        input: String,
     },
     #[structopt(name = "decode")]
-    Decode,
+    Decode {
+        /// The data need to be decoded.
+        #[structopt(short = "i", long = "input")]
+        input: String,
+    },
 }
 
-fn main() -> Result<(), ExitFailure> {
-    env_logger::init();
+fn main() -> Result<()> {
     let opts = Opts::from_args();
     match opts.mode {
-        Mode::Encode { base } => encode(base),
-        Mode::Decode => decode(),
+        Mode::Encode { base, input } => encode(base, input.as_bytes()),
+        Mode::Decode { input } => decode(&input),
     }
 }
 
 #[derive(Debug)]
 struct StrBase(Base);
 
-impl fmt::Display for StrBase {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+impl std::fmt::Display for StrBase {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
         let base_str = match self.0 {
+            Base::Identity => "identity",
             Base::Base2 => "base2",
             Base::Base8 => "base8",
             Base::Base10 => "base10",
-            Base::Base16Upper => "base16",
-            Base::Base16Lower => "base16-lower",
-            Base::Base32UpperNoPad => "base32",
-            Base::Base32UpperPad => "base32-pad",
-            Base::Base58flickr => "base58-flickr",
-            Base::Base58btc => "base58-btc",
-            Base::Base64UpperNoPad => "base64",
-            Base::Base64UpperPad => "base64-pad",
-            Base::Base64UrlUpperNoPad => "base64-url",
-            Base::Base64UrlUpperPad => "base64-url-pad",
+            Base::Base16Lower => "base16",
+            Base::Base16Upper => "base16upper",
+            Base::Base32HexLower => "base32hex",
+            Base::Base32HexUpper => "base32hexupper",
+            Base::Base32HexPadLower => "base32hexpad",
+            Base::Base32HexPadUpper => "base32hexpadupper",
+            Base::Base32Lower => "base32",
+            Base::Base32Upper => "base32upper",
+            Base::Base32PadLower => "base32pad",
+            Base::Base32PadUpper => "base32padupper",
+            Base::Base32Z => "base32z",
+            Base::Base58Flickr => "base58flickr",
+            Base::Base58Btc => "base58btc",
+            Base::Base64 => "base64",
+            Base::Base64Pad => "base64pad",
+            Base::Base64Url => "base64url",
+            Base::Base64UrlPad => "base64urlpad",
         };
         write!(f, "{}", base_str)
     }
 }
 
-impl FromStr for StrBase {
+impl std::str::FromStr for StrBase {
     type Err = Error;
 
     fn from_str(base_str: &str) -> Result<Self, Self::Err> {
         let base = match base_str {
+            "identity" => Ok(Base::Identity),
             "base2" => Ok(Base::Base2),
             "base8" => Ok(Base::Base8),
             "base10" => Ok(Base::Base10),
-            "base16" => Ok(Base::Base16Upper),
-            "base16-lower" => Ok(Base::Base16Lower),
-            "base32" => Ok(Base::Base32UpperNoPad),
-            "base32-pad" => Ok(Base::Base32UpperPad),
-            "base58-flickr" => Ok(Base::Base58flickr),
-            "base58-btc" => Ok(Base::Base58btc),
-            "base64" => Ok(Base::Base64UpperNoPad),
-            "base64-pad" => Ok(Base::Base64UpperPad),
-            "base64-url" => Ok(Base::Base64UrlUpperNoPad),
-            "base64-url-pad" => Ok(Base::Base64UrlUpperPad),
-            _ => Err(format_err!("Unknown base {:?}", base_str)),
+            "base16" => Ok(Base::Base16Lower),
+            "base16upper" => Ok(Base::Base16Upper),
+            "base32hex" => Ok(Base::Base32HexLower),
+            "base32hexupper" => Ok(Base::Base32HexUpper),
+            "base32hexpad" => Ok(Base::Base32HexPadLower),
+            "base32hexpadupper" => Ok(Base::Base32HexPadUpper),
+            "base32" => Ok(Base::Base32Lower),
+            "base32upper" => Ok(Base::Base32Upper),
+            "base32pad" => Ok(Base::Base32PadLower),
+            "base32padupper" => Ok(Base::Base32PadUpper),
+            "base32z" => Ok(Base::Base32Z),
+            "base58flickr" => Ok(Base::Base58Flickr),
+            "base58btc" => Ok(Base::Base58Btc),
+            "base64" => Ok(Base::Base64),
+            "base64pad" => Ok(Base::Base64Pad),
+            "base64url" => Ok(Base::Base64Url),
+            "base64urlpad" => Ok(Base::Base64UrlPad),
+            _ => return Err(anyhow!("Unknown base: {:?}", base_str)),
         };
         base.map(Self)
     }
@@ -87,26 +106,16 @@ impl From<StrBase> for Base {
     }
 }
 
-fn encode(base: StrBase) -> Result<(), ExitFailure> {
-    log::debug!("encoding with {}", base);
-    let mut stdin = io::stdin();
-    let mut stdout = io::stdout();
-    let mut buffer = Vec::new();
-    stdin.read_to_end(&mut buffer)?;
-    log::debug!("read {:?} from stdin", buffer);
-    let result = multibase::encode(base.into(), buffer.as_slice());
-    stdout.write_all(result.as_bytes())?;
+fn encode(base: StrBase, input: &[u8]) -> Result<()> {
+    println!("Encode {:?} with {}", input, base);
+    let result = multibase::encode(base.into(), input);
+    println!("Result: {}", result);
     Ok(())
 }
 
-fn decode() -> Result<(), ExitFailure> {
-    let mut stdin = io::stdin();
-    let mut stdout = io::stdout();
-    let mut buffer = String::new();
-    stdin.read_to_string(&mut buffer)?;
-    log::debug!("read {:?} from stdin", buffer);
-    let (base, result) = multibase::decode(buffer.as_str())?;
-    log::debug!("detected {}", StrBase(base));
-    stdout.write_all(result.as_slice())?;
+fn decode(input: &str) -> Result<()> {
+    println!("Decode {:?}", input);
+    let (base, result) = multibase::decode(input)?;
+    println!("Result: {}, {:?}", StrBase(base), result);
     Ok(())
 }