Signed-off-by: koushiro <koushiro.cqx@gmail.com>

* Implement all multibase encodings
* Refactor lib and cli

Close #15