@koushiro 2020-02-21
#Benchmark:

## Before

```
base_x                  time:   [350.14 us 358.97 us 368.72 us]
base32                  time:   [2.9558 us 2.9993 us 3.0583 us]
Found 5 outliers among 100 measurements (5.00%)
  4 (4.00%) high mild
  1 (1.00%) high severe
base64                  time:   [798.46 ns 803.70 ns 809.57 ns]
Found 3 outliers among 100 measurements (3.00%)
  2 (2.00%) high mild
  1 (1.00%) high severe

base_x #2               time:   [64.800 us 64.992 us 65.201 us]
base32 #2               time:   [2.9455 us 2.9594 us 2.9757 us]
Found 7 outliers among 100 measurements (7.00%)
  3 (3.00%) high mild
  4 (4.00%) high severe
base64 #2               time:   [807.48 ns 809.73 ns 812.26 ns]
Found 8 outliers among 100 measurements (8.00%)
  8 (8.00%) high mild
```

## After: 

```
encode/base32           time:   [904.17 ns 905.82 ns 908.09 ns]
Found 4 outliers among 100 measurements (4.00%)
  2 (2.00%) high mild
  2 (2.00%) high severe
encode/base58btc        time:   [331.32 us 331.44 us 331.59 us]
Found 3 outliers among 100 measurements (3.00%)
  1 (1.00%) high mild
  2 (2.00%) high severe
encode/base64           time:   [739.74 ns 740.53 ns 741.41 ns]
Found 2 outliers among 100 measurements (2.00%)
  1 (1.00%) high mild
  1 (1.00%) high severe

decode/base32           time:   [817.73 ns 818.51 ns 819.38 ns]
Found 6 outliers among 100 measurements (6.00%)
  4 (4.00%) high mild
  2 (2.00%) high severe
decode/base58btc        time:   [67.511 us 67.550 us 67.589 us]
Found 3 outliers among 100 measurements (3.00%)
  2 (2.00%) high mild
  1 (1.00%) high severe
decode/base64           time:   [775.68 ns 776.13 ns 776.62 ns]
Found 9 outliers among 100 measurements (9.00%)
  1 (1.00%) low mild
  2 (2.00%) high mild
  6 (6.00%) high severe
```